oracle should work

+++ lib/rucio/db/sqla/migrate_repo/versions/14ec5aeb64cf_add_request_external_host.py
@@ -1,40 +1,53 @@
-"""
-  Copyright European Organization for Nuclear Research (CERN)
+# Copyright 2013-2019 CERN for the benefit of the ATLAS collaboration.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# Authors:
+# - Mario Lassnig <mario.lassnig@cern.ch>, 2014-2019
+# - Vincent Garonne <vincent.garonne@cern.ch>, 2017
 
-  Licensed under the Apache License, Version 2.0 (the "License");
-  You may not use this file except in compliance with the License.
-  You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
-
-  Authors:
-  - Mario Lassnig, <mario.lassnig@cern.ch>, 2014
-  - Vincent Garonne, <vincent.garonne@cern.ch>, 2017
-
-add request external_host
-
-Revision ID: 14ec5aeb64cf
-Revises: 52fd9f4916fa
-Create Date: 2014-08-22 13:25:22.132950
-
-"""
-
-from alembic.op import add_column, drop_column
+''' add request external_host '''
 
 import sqlalchemy as sa
 
-# revision identifiers, used by Alembic.
-revision = '14ec5aeb64cf'  # pylint:disable=invalid-name
-down_revision = '52fd9f4916fa'  # pylint:disable=invalid-name
+from alembic import context
+from alembic.op import add_column, drop_column
+
+
+# Alembic revision identifiers
+revision = '14ec5aeb64cf'
+down_revision = '52fd9f4916fa'
 
 
 def upgrade():
     '''
-    upgrade method
+    Upgrade the database to this revision
     '''
-    add_column('requests', sa.Column('external_host', sa.String(256)))
+
+    if context.get_context().dialect.name in ['oracle', 'mysql']:
+        add_column('requests', sa.Column('external_host', sa.String(256)))
+
+    elif context.get_context().dialect.name == 'postgresql':
+        pass
 
 
 def downgrade():
     '''
-    downgrade method
+    Downgrade the database to the previous revision
     '''
-    drop_column('requests', 'external_host')
+
+    if context.get_context().dialect.name in ['oracle', 'mysql']:
+        drop_column('requests', 'external_host')
+
+    elif context.get_context().dialect.name == 'postgresql':
+        pass
