+++ lib/rucio/db/sqla/migrate_repo/versions/22cf51430c78_add_availability_column_to_table_rses.py
@@ -1,41 +1,50 @@
-"""
-  Copyright European Organization for Nuclear Research (CERN)
+# Copyright 2013-2019 CERN for the benefit of the ATLAS collaboration.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# Authors:
+# - Vincent Garonne <vincent.garonne@cern.ch>, 2014-2017
+# - Cedric Serfon <cedric.serfon@cern.ch>, 2014
+# - Mario Lassnig <mario.lassnig@cern.ch>, 2014-2019
 
-  Licensed under the Apache License, Version 2.0 (the "License");
-  You may not use this file except in compliance with the License.
-  You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
-
-  Authors:
-  - Vincent Garonne, <vincent.garonne@cern.ch>, 2014-2017
-  - Cedric Serfon, <cedric.serfon@cern.ch>, 2014
-  - Mario Lassnig, <mario.lassnig@cern.ch>, 2014
-
-Add availability column to table RSEs
-
-Revision ID: 22cf51430c78
-Revises: 49a21b4d4357
-Create Date: 2014-06-12 14:54:23.160946
-
-"""
-
-from alembic.op import add_column, drop_column
+''' add availability column to table RSEs '''
 
 import sqlalchemy as sa
 
-# revision identifiers, used by Alembic.
-revision = '22cf51430c78'  # pylint:disable=invalid-name
-down_revision = '49a21b4d4357'  # pylint:disable=invalid-name
+from alembic import context
+from alembic.op import add_column, drop_column
+
+
+# Alembic revision identifiers
+revision = '22cf51430c78'
+down_revision = '49a21b4d4357'
 
 
 def upgrade():
     '''
-    upgrade method
+    Upgrade the database to this revision
     '''
-    add_column('rses', sa.Column('availability', sa.Integer, server_default='7'))
+
+    if context.get_context().dialect.name in ['oracle', 'mysql']:
+        add_column('rses', sa.Column('availability', sa.Integer, server_default='7'))
+
+    elif context.get_context().dialect.name == 'postgresql':
+        pass
 
 
 def downgrade():
-    '''
-    downgrade method
-    '''
-    drop_column('rses', 'availability')
+    if context.get_context().dialect.name in ['oracle', 'mysql']:
+        drop_column('rses', 'availability')
+
+    elif context.get_context().dialect.name == 'postgresql':
+        pass
