+++ lib/rucio/db/sqla/migrate_repo/versions/44278720f774_update_requests_typ_sta_upd_idx_index.py
@@ -1,41 +1,56 @@
-# Copyright European Organization for Nuclear Research (CERN)
+# Copyright 2013-2019 CERN for the benefit of the ATLAS collaboration.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
-# You may not use this file except in compliance with the License.
-# You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 #
 # Authors:
-# - Wen Guan, <wen.guan@cern.ch>, 2015
-# - Vincent Garonne, <vincent.garonne@cern.ch>, 2017
+# - Wen Guan <wen.guan@cern.ch>, 2015
+# - Vincent Garonne <vincent.garonne@cern.ch>, 2017
+# - Mario Lassnig <mario.lassnig@cern.ch>, 2019
 
+''' update REQUESTS_TYP_STA_UPD_IDX index '''
 
-"""update REQUESTS_TYP_STA_UPD_IDX index
-
-Revision ID: 44278720f774
-Revises: 40ad39ce3160
-Create Date: 2015-05-06 16:37:06.805315
-
-"""
+from alembic import context
 from alembic.op import create_index, drop_index
 
-# revision identifiers, used by Alembic.
+
+# Alembic revision identifiers
 revision = '44278720f774'
 down_revision = '40ad39ce3160'
 
 
 def upgrade():
     '''
-    upgrade method
+    Upgrade the database to this revision
     '''
-    drop_index('REQUESTS_TYP_STA_UPD_IDX', 'requests')
-    create_index('REQUESTS_TYP_STA_UPD_IDX', 'requests', ['request_type', 'state', 'activity'])
-    create_index('REQUESTS_TYP_STA_UPD_IDX_OLD', 'requests', ['request_type', 'state', 'updated_at'])
+
+    if context.get_context().dialect.name in ['oracle', 'mysql']:
+        drop_index('REQUESTS_TYP_STA_UPD_IDX', 'requests')
+        create_index('REQUESTS_TYP_STA_UPD_IDX', 'requests', ['request_type', 'state', 'activity'])
+        create_index('REQUESTS_TYP_STA_UPD_IDX_OLD', 'requests', ['request_type', 'state', 'updated_at'])
+
+    elif context.get_context().dialect.name == 'postgresql':
+        pass
 
 
 def downgrade():
     '''
-    downgrade method
+    Downgrade the database to the previous revision
     '''
-    drop_index('REQUESTS_TYP_STA_UPD_IDX', 'requests')
-    drop_index('REQUESTS_TYP_STA_UPD_IDX_OLD', 'requests')
-    create_index('REQUESTS_TYP_STA_UPD_IDX', 'requests', ['request_type', 'state', 'updated_at'])
+
+    if context.get_context().dialect.name in ['oracle', 'mysql']:
+        drop_index('REQUESTS_TYP_STA_UPD_IDX', 'requests')
+        drop_index('REQUESTS_TYP_STA_UPD_IDX_OLD', 'requests')
+        create_index('REQUESTS_TYP_STA_UPD_IDX', 'requests', ['request_type', 'state', 'updated_at'])
+
+    elif context.get_context().dialect.name == 'postgresql':
+        pass
